+++ jvm-streaming/tests/Main.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE OverloadedStrings #-}
 module Main where
 
 import Language.Java (withJVM)
@@ -5,4 +6,6 @@ import qualified Spec
 import Test.Hspec
 
 main :: IO ()
-main = withJVM [] $ hspec Spec.spec
+main =
+    withJVM ["-Djava.class.path=../jvm-batching/build/libs/jvm-batching.jar"] $
+      hspec Spec.spec