+++ jvm-streaming/CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Next release
 
+## Added
+
+* benchmarks for batched marshalling of streams.
+
 ### Changed
 
+* jvm-streaming: Implement marshalling of streams with jvm-batching.
 * Store lookahead element in the Java side when reflecting streams.
 
 ## [0.2.1] - 2017-08-31